+++ apps/web/src/api/apiRequest.client.ts
@@ -1,4 +1,6 @@
-import { getRefreshToken } from '@/auth/apis/refresh';
+'use client';
+
+import { getRefreshToken } from '@/auth/apis/refresh.client';
 import { getAccessToken } from '@/auth/token.client';
 import { createApiRequest } from '@/api/apiRequest.base';
