+++ apps/web/src/api/apiRequest.server.ts
@@ -1,6 +1,6 @@
 import 'server-only';
 
-import { getRefreshToken } from '@/auth/apis/refresh';
+import { getRefreshToken } from '@/auth/apis/refresh.server';
 import { getAccessToken } from '@/auth/token.server';
 import { createApiRequest } from '@/api/apiRequest.base';
