+++ apps/web/src/auth/apis/refresh.server.ts
@@ -1,0 +1,20 @@
+import 'server-only';
+
+import { cookies } from 'next/headers';
+import { SERVER_API_BASE_URL } from '@/api/constants/api';
+
+export const getRefreshToken = async () => {
+  const cookieStore = await cookies();
+  const cookieHeader = cookieStore
+    .getAll()
+    .map(({ name, value }) => `${name}=${encodeURIComponent(value)}`)
+    .join('; ');
+
+  return fetch(`${SERVER_API_BASE_URL}/api/v1/auth/reissue`, {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+      ...(cookieHeader ? { Cookie: cookieHeader } : {}),
+    },
+  });
+};
